[FIX] Improve Wine not available / not found handler (#2327)
I am trying to find a solution to this issue people complain sometimes
when wine was not found or maybe it was deleted somehow, the game fails
to launch and you need to go to the settings and change it back and
forth, etc.
So my solution is pretty simple:
- On launch, if the game is not native, check if the wine is valid. This
means that not only the needed wine binary is available but also
wineserver and wineboot for type = wine.
- if not, show a dialog to the user proposing another wine version.
- If the one set as default is not valid somehow, query the list of
wine, proton, crossover, etc. And get the first one.
- if none was found, then it will fail and show a message to the user.
- To avoid showing two dialogs, Heroic first checks the default one, if
it fails it will check any wine version. If it works then it shows the
dialog.

Besides that, I added a few QOL improvements on the settings part.
I mean, we have a WHOLE class to deal with the settings and there is no
way to SET a Single setting, everytime we change a setting in Heroic it
rewrites the whole file again. So if we need to change a setting on the
backend we need to query the whole list, then change the key inside that
object, and then rewrite the `settings` again. That is not practical.

So I added the `setSetting` method to the class on the backend and also
added the ipc call on the frontend. so we can now set one single setting
on backend and frontend, making our life easier if we need to handle
something on runtime.

Another small method I added was `pathExists`, so we can query on the
frontend if a path exists and use it to update the UI or show some
message to the user, etc.

On Wine settings:

![image](https://user-images.githubusercontent.com/26871415/211602830-ee89b246-9090-4435-8f9e-1b3b0b6b811b.png)

On Launching:

![image](https://user-images.githubusercontent.com/26871415/211602926-d3eac9e1-386d-4a4a-805f-a542f2f2529d.png)

---

Use the following Checklist if you have changed something on the Backend
or Frontend:

- [x] Tested the feature and it's working on a current and clean
install.
- [ ] Tested the main App features and they are still working on a
current and clean install. (Login, Install, Play, Uninstall, Move games,
etc.)
- [ ] Created / Updated Tests (If necessary)
- [ ] Created / Updated documentation (If necessary)