[Feat] Add settings to Auto Update games and Ignore Game Updates (#2308)
This adds a new global setting to auto update games, so on the game card
it checks if the game has update and if the setting is enabled and if
the ignore game updates is disabled. Then it send the game to the queue.

I think it is redundant to add this code to the game page since the
library is where we check for the game updates so I am checking only on
the game card.

Another option that was added was to ignore game updates. This Will
ignore the auto update feature and also when clicking play it won't show
the update warning.

I did not add this setting to the global game settings since I think
this should be set individually.

---

Use the following Checklist if you have changed something on the Backend
or Frontend:

- [x] Tested the feature and it's working on a current and clean
install.
- [ ] Tested the main App features and they are still working on a
current and clean install. (Login, Install, Play, Uninstall, Move games,
etc.)
- [ ] Created / Updated Tests (If necessary)
- [ ] Created / Updated documentation (If necessary)