[Ref/Feat] System Information (#3027)
* Move `getLegendaryVersion`, `getGogdlVersion`, `getNileVersion` and
`getHericVersion` outside of utils.ts

The behavior of everything but `getHeroicVersion` stayed the same.
Heroic's own version codenames will now live inside `package.json` (right next
to the version number) for convenience

* Move `searchForExecutableOnPath` out of utils.ts

This also adds the `hasExecutable` shorthand

* Add a generic "download this file" utility

We've had variations of this in several smaller `util` files, let's go for
making one generic one instead

* Add a generic "run this process" utility

Versions of something like this (`callRunner` for example) have existed as
purpose-built functions, but implementing something new that intents to spawn
processes always needed a new "spawn this process with these arguments and give
me the output" function along with it. Time to change that

* Add a new `systeminfo` utility

This is essentially the `systeminformation` module, but tailor-made for our use
case (no longer getting stuck, gathering more information, respecting
Flatpak paths)

* Add a new "System Information" section to the Settings

This offers the user's system information in a more visual and easy-to-copy way

* Add an Apple logo (for M1/M2/... CPUs)

* Report correct memory use on Linux & macOS

* Add a property describing whether we're in a Flatpak container